Switch options and mode bits to uint32_t from unsigned long.
This is a bitmask, so the number of bits available should be the same
across all platforms.

Reviewed-on: https://boringssl-review.googlesource.com/4556

+++ include/openssl/ssl.h
@@ -837,8 +837,8 @@ struct ssl_ctx_st {
   /* Default values to use in SSL structures follow (these are copied by
    * SSL_new) */
 
-  unsigned long options;
-  unsigned long mode;
+  uint32_t options;
+  uint32_t mode;
   long max_cert_list;
 
   struct cert_st /* CERT */ *cert;
@@ -1340,8 +1340,8 @@ struct ssl_st {
   /* for server side, keep the list of CA_dn we can use */
   STACK_OF(X509_NAME) *client_CA;
 
-  unsigned long options; /* protocol behaviour */
-  unsigned long mode;    /* API behaviour */
+  uint32_t options; /* protocol behaviour */
+  uint32_t mode;    /* API behaviour */
   long max_cert_list;
   int client_version; /* what was passed, used for
                        * SSLv3/TLS rollback check */
